create and emit ParamDeprecationWarning
